requirements.txt
fastapi>=0.115.6
pydantic>=2.0.0
mangum>=0.19.0
requests>=2.32.3
python-multipart>=0.0.9
azure-ai-vision-face>=1.0.0b2
